generating initial cc-config and -interface submodels

+++ modules/de.dfki.cos.basys.controlcomponent.server/src/main/java/de/dfki/cos/basys/controlcomponent/server/opcua/ControlComponentNamespace.java
@@ -13,6 +13,17 @@ package de.dfki.cos.basys.controlcomponent.server.opcua;
 import java.util.List;
 import java.util.Optional;
 
+import org.eclipse.basyx.aas.manager.ConnectedAssetAdministrationShellManager;
+import org.eclipse.basyx.submodel.metamodel.api.identifier.IdentifierType;
+import org.eclipse.basyx.submodel.metamodel.api.qualifier.haskind.ModelingKind;
+import org.eclipse.basyx.submodel.metamodel.api.reference.enums.KeyElements;
+import org.eclipse.basyx.submodel.metamodel.map.SubModel;
+import org.eclipse.basyx.submodel.metamodel.map.qualifier.LangStrings;
+import org.eclipse.basyx.submodel.metamodel.map.reference.Key;
+import org.eclipse.basyx.submodel.metamodel.map.submodelelement.SubmodelElementCollection;
+import org.eclipse.basyx.submodel.metamodel.map.submodelelement.dataelement.property.Property;
+import org.eclipse.basyx.submodel.metamodel.map.submodelelement.dataelement.property.valuetypedef.PropertyValueTypeDef;
+import org.eclipse.basyx.vab.modelprovider.lambda.VABLambdaProviderHelper;
 import org.eclipse.milo.opcua.sdk.core.Reference;
 import org.eclipse.milo.opcua.sdk.server.OpcUaServer;
 import org.eclipse.milo.opcua.sdk.server.api.DataItem;
@@ -25,11 +36,14 @@ import org.eclipse.milo.opcua.sdk.server.util.SubscriptionModel;
 import org.eclipse.milo.opcua.stack.core.Identifiers;
 import org.eclipse.milo.opcua.stack.core.types.builtin.LocalizedText;
 import org.eclipse.milo.opcua.stack.core.types.builtin.NodeId;
+import org.eclipse.milo.opcua.stack.core.types.structured.EndpointDescription;
+import org.eclipse.milo.opcua.stack.server.EndpointConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.eventbus.Subscribe;
 
+import de.dfki.cos.basys.aas.component.AasComponentContext;
 import de.dfki.cos.basys.common.component.Component;
 import de.dfki.cos.basys.common.component.ComponentContext;
 import de.dfki.cos.basys.common.component.manager.impl.ComponentManagerEvent;
@@ -55,7 +69,7 @@ public class ControlComponentNamespace extends ManagedNamespace {
     @Override
     protected void onShutdown() {
     	super.onShutdown();    
-    	ComponentContext.getStaticContext().getEventBus().unregister(this);		
+    	AasComponentContext.getStaticContext().getEventBus().unregister(this);		
     }
     
     @Override
@@ -80,7 +94,7 @@ public class ControlComponentNamespace extends ManagedNamespace {
             .map(UaVariableNode.class::cast)
             .forEach(n -> n.setMinimumSamplingInterval(100.0));
     	
-    	ComponentContext.getStaticContext().getEventBus().register(this); 
+    	AasComponentContext.getStaticContext().getEventBus().register(this); 
     }
     
     @Override
@@ -108,16 +122,31 @@ public class ControlComponentNamespace extends ManagedNamespace {
 		if (ev.getType() == Type.COMPONENT_ADDED) {
 			Component component = ev.getComponent();
 			if (component instanceof ControlComponent) {
-				UaNode node = createComponentRootNode((ControlComponent) component);
+				ControlComponent cc = (ControlComponent) component;
+				UaNode node = createComponentRootNode(cc);
 				
 				// Make sure our new folder shows up under the server's Objects folder.
 				node.addReference(new Reference(node.getNodeId(), Identifiers.Organizes,
 						Identifiers.ObjectsFolder.expanded(), false));
+				
+
+				//create and register config submodel
+				ConnectedAssetAdministrationShellManager aasManager = new ConnectedAssetAdministrationShellManager(AasComponentContext.getStaticContext().getAasRegistry());
+				SubModel configSubmodel = createConfigSubmodel(node, cc);
+				aasManager.createSubModel(cc.getAasId(), configSubmodel);			
+				
 			}
 		}
 		else if (ev.getType() == Type.COMPONENT_DELETED) {
-			Optional<UaNode> node = getNodeManager().removeNode(newNodeId(ev.getValue()));
-			node.ifPresent(n -> n.delete());
+			Component component = ev.getComponent();
+			if (component instanceof ControlComponent) {
+				ControlComponent cc = (ControlComponent) component;
+				Optional<UaNode> node = getNodeManager().removeNode(newNodeId(ev.getValue()));
+				node.ifPresent(n -> n.delete());	
+				
+				//TODO: remove and unregister config sm				
+			}
+			
 		}
 	}
     
@@ -136,6 +165,73 @@ public class ControlComponentNamespace extends ManagedNamespace {
     
     private NodeId newNodeId() {
     	return newNodeId(nodeIndex++);
+    }
+    
+    private SubModel createConfigSubmodel(UaNode node, ControlComponent component) {
+    	SubModel submodel = new SubModel();		
+		submodel.setIdShort(component.getId() + "-ccc");
+		submodel.setIdentification(component.getSubmodelId().getIdType(), component.getSubmodelId().getId().replace("control-component", "control-component-config"));
+		submodel.setModelingKind(ModelingKind.INSTANCE);
+		submodel.setDescription(new LangStrings("en-US", "ControlComponent configuration submodel for component " + component.getId()));
+		submodel.setSemanticId(new org.eclipse.basyx.submodel.metamodel.map.reference.Reference(new Key(KeyElements.CONCEPTDESCRIPTION, false, "ControlComponentConfiguration", IdentifierType.CUSTOM)));
+		
+//		SubmodelElementCollection status = new SubmodelElementCollection();
+//		status.setIdShort("STATUS");
+//		submodel.addSubModelElement(status);
+			
+		Property nodeId = new Property();
+		nodeId.setIdShort("nodeId");
+		nodeId.set(node.getNodeId().toParseableString());
+		submodel.addSubModelElement(nodeId);
+		
+//		for (EndpointConfiguration ec : this.getServer().getConfig().getEndpoints()) {
+//			SubmodelElementCollection endpointConfig = new SubmodelElementCollection();
+//			endpointConfig.setIdShort("endpoint");
+//
+//			Property endpoint = new Property();
+//			endpoint.setIdShort("endpoint");
+//			endpoint.set(ed.getEndpointUrl());
+//			
+//			Property transportProfile = new Property();
+//			transportProfile.setIdShort("transportProfile");
+//			transportProfile.set(ed.getTransportProfileUri());
+//
+//			Property securityPolicy = new Property();
+//			securityPolicy.setIdShort("securityPolicy");
+//			securityPolicy.set(ed.getSecurityPolicyUri());			
+//
+//			endpointConfig.addElement(endpoint);
+//			endpointConfig.addElement(transportProfile);
+//			endpointConfig.addElement(securityPolicy);
+//
+//			submodel.addSubModelElement(endpointCollection);
+//		}
+				
+		for (EndpointDescription ed : this.getServer().getEndpointDescriptions()) {
+			
+			SubmodelElementCollection endpointCollection = new SubmodelElementCollection();
+			endpointCollection.setIdShort("endpoint");
+
+			Property endpoint = new Property();
+			endpoint.setIdShort("endpoint");
+			endpoint.set(ed.getEndpointUrl());
+			
+			Property transportProfile = new Property();
+			transportProfile.setIdShort("transportProfile");
+			transportProfile.set(ed.getTransportProfileUri());
+
+			Property securityPolicy = new Property();
+			securityPolicy.setIdShort("securityPolicy");
+			securityPolicy.set(ed.getSecurityPolicyUri());			
+
+			endpointCollection.addElement(endpoint);
+			endpointCollection.addElement(transportProfile);
+			endpointCollection.addElement(securityPolicy);
+
+			submodel.addSubModelElement(endpointCollection);
+		}		
+		
+		return submodel;
     }
     
 }
